+++ Assets/Prefabs/Minion/Slime/FireflySlime.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 061f5cea81afe4666911576036903dd4
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
