+++ Assets/Scripts/Creature/Minion/Slime/Components.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5f6d7b441685d4b8888c0414f1b06d13
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
