+++ Assets/Scripts/Creature/Minion/Slime/Components/SlimeAnimator.cs
@@ -19,6 +19,11 @@ namespace Creature.Minion.Slime
             animator.sprites = sprites;
             return animator;
         }
+        
+        public void SetSprites(SlimeSprites sprites)
+        {
+            this.sprites = sprites;
+        }
 
         public enum AnimationType
         {
@@ -40,16 +45,17 @@ namespace Creature.Minion.Slime
     
         public void AsyncPlay(AnimationType type, bool loop = false)
         {
-            if (currentAnimation != null)
-            {
-                Stop();
-            }
+            Stop();
             currentAnimation = StartCoroutine(SyncPlay(type, loop));
         }
     
         public void Stop()
         {
-            StopCoroutine(currentAnimation);
+            if (currentAnimation != null)
+            {
+                StopCoroutine(currentAnimation);
+            }
+
             currentAnimation = null;
         }
