+++ Assets/Scripts/Creature/Minion/Slime/Components/SlimeSprites.cs.meta
