+++ Assets/Scripts/Creature/Minion/Slime/FireflySlime.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4f40d890f86b3457e835b60ddf1e9af6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
