+++ Assets/Scripts/Creature/Minion/Spawner.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9b44848eecb314c219801cc212430ea9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
