+++ Assets/Scripts/Creature/Minion/Spawner/FireflySlimeSpawner.cs
@@ -1,0 +1,7 @@
+namespace Creature.Minion.Slime
+{
+    public class FireflySlimeSpawner : MobSpawner<FireflySlime>
+    {
+        
+    }
+}
