+++ Assets/Scripts/Creature/Minion/Spawner/FireflySlimeSpawner.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 500d713e1bc6407cb770f44e0bb507a2
+timeCreated: 1741407221
