+++ Assets/Scripts/Creature/Minion/Spawner/SlimeSpawner.cs
@@ -5,3 +5,4 @@ namespace Creature.Minion.Slime
         
     }
 }
+
